Moves functions from scripts.js to domUpdates methods
